+++ packages/article/src/__tests__/android/related-article.android.test.js
@@ -1,5 +1,0 @@
-import relatedArticlesShared from "../related-articles-shared";
-
-describe("Related articles tests on android", () => {
-  relatedArticlesShared();
-});