+++ packages/card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.25.6"></a>
+## [0.25.6](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.5...@times-components/card@0.25.6) (2018-04-05)
+
+
+
+
+**Note:** Version bump only for package @times-components/card
+
 <a name="0.25.5"></a>
 ## [0.25.5](https://github.com/newsuk/times-components/compare/@times-components/card@0.25.4...@times-components/card@0.25.5) (2018-04-05)